% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident for given state in a given year
This function plots the accident in a given state for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{unique}{state number - numeric}

\item{valid}{year - numeric}
}
\value{
1. plotted map of accident for the year
2. "invalid state" if state number not found
3. "No accident to plot " if for a given year and given state no accident occurred.
}
\description{
Plot accident for given state in a given year
This function plots the accident in a given state for a given year.
}
\examples{
setwd(system.file("extdata", package = "FarsProject"))
fars_map_state(050,2014)

}
