% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file of Fatality Analysis and Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{to be printed}
}
\value{
returns the dataframe from read file.
}
\description{
This function reads the file as csv as a dataframe and prompt error if file does not existi
}
\examples{
setwd(system.file("extdata", package = "FarsProject"))
fars_read('accident_2013.csv.bz2')

}
