% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years data from Fatality Analysis and Reporting system}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{- year for which data need to be extracted}
}
\value{
returns the dataframe from read file.
}
\description{
Reads the month and year data from the file
}
