% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the FARS data for a given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{-year for which summarization need to be done}
}
\value{
return N X 2 dataframe with month and number of accident
}
\description{
It summarizes the monthly accident in a given year
}
\examples{
setwd(system.file("extdata", package = "FarsProject"))
fars_summarize_years(2013)

}
