% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create the FARS filename using year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{to be embeded in the filename}
}
\value{
print the new file name with year from parameter added to it.
}
\description{
It will make the filename with year embeded in its name
}
