% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw map of NHTSA FARS accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{USA state number, numeric.}

\item{year}{Year of ovservations, numeric.}
}
\value{
No return value (NULL).
}
\description{
Reads Fatality Analysis Reporting System (FARS) data files and plots maps of accidents for the specified state and year
with the graphics package.
}
\details{
Will stop with error if given state number is invalid. Will warn if no data are available for a given year.
}
\examples{
\dontrun{
fars_map_state(1, 2015)
}

}
