% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Get the name of NHTSA FARS data file for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year of observations.}
}
\value{
A character vector containing the name of the file, without full path.
}
\description{
Makes a filename, without the full path, of a Fatality Analysis Reporting System (FARS) data file
in CSV format for a specified year.
}
\examples{
\dontrun{
make_filename(2015)
}

}
